fix QA 4

+++ app/Administration/Role/Routes/api.php
@@ -3,14 +3,12 @@
 use App\Administration\Role\Controllers\RoleController;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['role:Super Admin'])->group(function (): void {
-    Route::controller(RoleController::class)->group(function (): void {
-        Route::get('/roles/permissions', 'permissionsIndex');
-        Route::post('/roles', 'create');
-        Route::patch('/roles/{role}', 'update');
-        Route::delete('/roles/{role}', 'delete');
-        Route::get('/roles', 'getAll');
-        Route::get('/roles/{role}', 'get');
-        Route::post('/roles/{role}/sync-permissions', 'syncPermissions');
-    });
+Route::controller(RoleController::class)->group(function (): void {
+    Route::get('/roles/permissions', 'permissionsIndex')->middleware('permission:role.permissionsIndex');
+    Route::post('/roles', 'create')->middleware('permission:role.create');
+    Route::patch('/roles/{role}', 'update')->middleware('permission:role.update');
+    Route::delete('/roles/{role}', 'delete')->middleware('permission:role.delete');
+    Route::get('/roles', 'getAll')->middleware('permission:role.getAll');
+    Route::get('/roles/{role}', 'get')->middleware('permission:role.get');
+    Route::post('/roles/{role}/sync-permissions', 'syncPermissions')->middleware('permission:role.syncPermissions');
 });
