+++ app/Administration/Tenant/Routes/api.php
@@ -3,12 +3,10 @@
 use App\Administration\Tenant\Controllers\TenantController;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['role:Super Admin'])->group(function (): void {
-    Route::controller(TenantController::class)->group(function (): void {
-        Route::post('/tenants', 'create');
-        Route::patch('/tenants/{tenant}', 'update');
-        Route::delete('/tenants/{tenant}', 'delete');
-        Route::get('/tenants', 'getAll');
-        Route::get('/tenants/{tenant}', 'get');
-    });
+Route::controller(TenantController::class)->group(function (): void {
+    Route::post('/tenants', 'create')->middleware('permission:tenant.create');
+    Route::patch('/tenants/{tenant}', 'update')->middleware('permission:tenant.update');
+    Route::delete('/tenants/{tenant}', 'delete')->middleware('permission:tenant.delete');
+    Route::get('/tenants', 'getAll')->middleware('permission:tenant.getAll');
+    Route::get('/tenants/{tenant}', 'get')->middleware('permission:tenant.get');
 });
