+++ app/Directory/Team/Routes/api.php
@@ -5,22 +5,22 @@ use App\Directory\Team\Controllers\PaymentController;
 use App\Directory\Team\Controllers\TeamController;
 use Illuminate\Support\Facades\Route;
 
-Route::controller(TeamController::class)->group(function () {
-    Route::post('/teams', 'create');
-    Route::patch('/teams/{team}', 'update');
-    Route::delete('/teams/{team}', 'delete');
-    Route::get('/teams', 'getAll');
-    Route::get('/teams/{team}', 'get');
+Route::controller(TeamController::class)->group(function (): void {
+    Route::post('/teams', 'create')->middleware('permission:team.create');
+    Route::patch('/teams/{team}', 'update')->middleware('permission:team.update');
+    Route::delete('/teams/{team}', 'delete')->middleware('permission:team.delete');
+    Route::get('/teams', 'getAll')->middleware('permission:team.getAll');
+    Route::get('/teams/{team}', 'get')->middleware('permission:team.get');
 });
 
-Route::controller(AttendanceController::class)->group(function () {
-    Route::get('/attendance/daily-summary', 'getDailySummary');
-    Route::get('/attendance/{teamId}', 'getByMonth');
-    Route::post('/attendance', 'store');
+Route::controller(AttendanceController::class)->group(function (): void {
+    Route::get('/attendance/daily-summary', 'getDailySummary')->middleware('permission:attendance.getDailySummary');
+    Route::get('/attendance/{teamId}', 'getByMonth')->middleware('permission:attendance.getByMonth');
+    Route::post('/attendance', 'store')->middleware('permission:attendance.store');
 });
 
-Route::controller(PaymentController::class)->group(function () {
-    Route::get('/payments/payroll', 'getPayroll');
-    Route::get('/payments', 'getByMonth');
-    Route::post('/payments', 'store');
+Route::controller(PaymentController::class)->group(function (): void {
+    Route::get('/payments/payroll', 'getPayroll')->middleware('permission:payment.getByMonth');
+    Route::get('/payments', 'getByMonth')->middleware('permission:payment.getByMonth');
+    Route::post('/payments', 'store')->middleware('permission:payment.store');
 });
