+++ app/Finance/Order/Routes/api.php
@@ -3,10 +3,10 @@
 use App\Finance\Order\Controllers\OrderController;
 use Illuminate\Support\Facades\Route;
 
-Route::controller(OrderController::class)->group(function () {
-    Route::post('/orders', 'create');
-    Route::patch('/orders/{order}', 'update');
-    Route::delete('/orders/{order}', 'delete');
-    Route::get('/orders', 'getAll');
-    Route::get('/orders/{order}', 'get');
+Route::controller(OrderController::class)->group(function (): void {
+    Route::post('/orders', 'create')->middleware('permission:order.create');
+    Route::patch('/orders/{order}', 'update')->middleware('permission:order.update');
+    Route::delete('/orders/{order}', 'delete')->middleware('permission:order.delete');
+    Route::get('/orders', 'getAll')->middleware('permission:order.getAll');
+    Route::get('/orders/{order}', 'get')->middleware('permission:order.get');
 });
