+++ app/Shared/Image/Routes/api.php
@@ -3,9 +3,9 @@
 use App\Shared\Image\Controllers\ImageController;
 use Illuminate\Support\Facades\Route;
 
-Route::controller(ImageController::class)->group(function() {
-    Route::post('/images', 'create');
-    Route::delete('/images/{image}', 'delete');
-    Route::get('/images', 'getAll');
-    Route::get('/images/{image}', 'get');
+Route::controller(ImageController::class)->group(function (): void {
+    Route::post('/images', 'create')->middleware('permission:image.create');
+    Route::delete('/images/{image}', 'delete')->middleware('permission:image.delete');
+    Route::get('/images', 'getAll')->middleware('permission:image.getAll');
+    Route::get('/images/{image}', 'get')->middleware('permission:image.get');
 });
